Add function to open a file asynchronously (#1342)
The function `open_file` in windows.py returns a Promise that shall be
resolved once the file is fully loaded. It's only safe to call `open_file`
from the main thread because the `on_load` event callback runs on the main
thread.

It should be used for applying workspace edits and renames.