Define a closure inside a closure
When defining a closure, care must be taken when defining it inside of an
infinite "while True" loop. Because a nonlocal reference in the closure
being used inside of the closure, can mutate outside of the closure.

To fix this, we define another closure around it that explicitly captures
the variable that we want to use in an input argument.