Run TransportCallbacks on ST's worker thread (#1348)
Previously, we ran the callbacks in the reader threads. We now run them
serialized through ST's worker thread. This saves us from some boring
boilerplate code to ensure callbacks are running serially.

When defining a closure, care must be taken when defining it inside of an
infinite "while True" loop. Because a nonlocal reference in the closure
being used inside of the closure, can mutate outside of the closure.

To fix this, we define another closure around it that explicitly captures
the variable that we want to use in an input argument. We do that with
functools.partial.

Co-authored-by: Rafał Chłodnicki <rchl2k@gmail.com>