Expose languageId on the params sent to on_workspace_configuration()
To fix https://github.com/sublimelsp/LSP-eslint/issues/36 I need to have
access to the "languageId" of the file that the server is requesting
configuration for.

This changes the "params" object so its schema will no longer match
the LSP spec but then it was nowhere stated that the params passed to
on_workspace_configuration() were matching the spec.