Replace version incompatibility dialog with a console message (#1376)
The latest version of Package Control is triggering an update of packages
on updating to the new ST version so there is no point in showing this
message now.

Keep it as a console message thought in case PC doesn't update for some
reason and the user reports a problem. That will make it easier to diagnose.