New server configuration key: auto_complete_selector
The auto_complete_selector is used to fine-tune the behavior of the AC popping
up. This selector is added to an entry of the "auto_complete_triggers" of the
view, alongside the trigger characters advertised by the language server.

When both the selector and at least one trigger matches, the AC is shown.

Note that we have to notify all SessionViews of new capability registrations,
so that they can update their trigger chars.