Fix Config tests
The tests for Config relied on a notion of equality. They implicitly relied
on equality of `id` (memory addresses). But now that we always create a new
Config in the WindowManager whether there is an override or not, this did
not work any longer. To solve this, we define what equality means.