Re-read workspace folders on starting a server (#1395)
This is a workaround for a case when user changes workspace folders
manually for which ST provides no notification. We'll just re-read the
window.folders() every time the user opens any view. If there were removed
or added folders we notify the appropriate servers that are interested
in this notification.

Resolves #1374