Run the text command from the correct view for "subl:" links
When the hover popup appears on an unfocused view, hyperlinks which point to
a text command via the `subl:` protocol run in another, focused, view. This is
undesired behavior. To fix this, focus the view just in time.

The is_enabled workarounds do not work because ST never calls is_enabled with
a `point` argument (only with an `event` argument).

Fixes #1313