Send back whole action object on choosing action in showMessageRequest
MessageActionItem is specified as only containing the "title" property [1].
Yet, vscode-eslint sends an object with "title" and "id" properties and
expects to get that "id" back when user chooses an action.

Since it's not clear what is expected I'm for now changing behavior to
send back original object. This "spec issue" is discussed in more detail at [2].

[1] https://microsoft.github.io/language-server-protocol/specification#window_showMessageRequest
[2] https://github.com/microsoft/language-server-protocol/issues/230