New server configuration key: auto_complete_selector (#1408)
The auto_complete_selector is used to fine-tune the behavior of the AC popping
up. This selector is added to an entry of the "auto_complete_triggers" of the
view, alongside the trigger characters advertised by the language server.

When both the selector and at least one trigger matches, the AC is shown.

Note that we have to notify all SessionViews of new capability registrations,
so that they can update their trigger chars. Fallback to the auto_complete_selector
of the view if nothing is specified Allow disabling the completion triggers from the
server entirely. You can also disable the trigger chars entirely with the boolean setting
ignore_server_trigger_chars.