Advertise support for regular expression syntax "ECMAScript"
This is only used for completions at the moment. With respect to completions,
the regex flavor is actually Perl, but Perl is a strict subset of ECMAScript.

Because the spec only lists one "well-known" flavor, we restrict ourselves to
this subset to achieve maximum compatibility with language servers that may
check for this particular value.

I have intentionally left out the version because:

- I suspect language servers will not check this value rigorously, and
- I don't actually know what version of ECMAScript corresponds to ST's internal
  implementation.