Make renaming more careful
When a server returns text edits for textDocument/rename that span more than
one file, we'll now show a dialog asking the user to confirm that he/she really
wants to apply the changes. The user can also select "Dry Run" to show the
locations of the text edits that will be applied in an output panel.

Resolves #873