Add AbstractPlugin.on_settings_changed API
This replaces the old "on_workspace_did_change_configuration" with
something more generic that will be called whenever client settings
change (currently those cannot really change during the lifetime of the
Session but might in the future).

The Plugin instance receives a reference to the object that lives as
long as the session and is also **the** configuration that is sent to
the server in other cases like when responding to
"workspace/configuration".

Relates to https://github.com/sublimelsp/LSP-pyright/issues/30