Fix: document highlights cannot be sent earlier than didChange
The text changes are buffered and sent after a 500ms timeout, provided that
nothing interesting happens like a trigger character.

The highlights are now sent after a 300ms timeout. This can cause the request
params to contain a document position that doesn't exist from the viewpoint of
the server, because those changes did not yet arrive (they will arrive another
200ms later).

This can be fixed by setting the timeout for highlights to 500ms.

Found in general rust-analyzer debugging.