Fix: document highlights cannot be sent earlier than didChange (#1507)
The text changes are buffered and sent after a 500ms timeout, provided that
nothing interesting happens like a trigger character.

The highlights are now sent after a 300ms timeout. This can cause the request
params to contain a document position that doesn't exist from the viewpoint of
the server, because those changes did not yet arrive (they will arrive another
200ms later).

This is fixed by setting all timeouts to 300ms for
textDocument{didChange,documentHighlight,codeAction}.

Found in general rust-analyzer debugging.