Path mapping Part II
This adds a ClientConfig argument to most of the functions in views.py so that
we can translate from local paths to remote paths.

Some functions are used by helper packages so we have to put the argument at the
end and make it optional. Version 1.3 should make them mandatory.