Never close the diags panel automatically, and more (#1525)
Because the diags panel is never closed automatically, it makes little sense
to have it automatically open whenever there's an error/warning. So the
setting

    auto_show_diagnostics_panel

is removed. Instead

    auto_show_diagnostics_panel_level

is used to determine when to automatically open the panel *on save*. Put the
value `0` here to never open the panel.

I've also added a neutral "idle" message for when there are no diagnostics.