Add A File Icon alias scopes to language-ids
This commit adds all alias scopes of A File Icon, which point to source
or text files.

Those aliases are created to attach unique icons to special named files,
such as unittests or makefiles. All those alias files either use
text.plain or another scope of a syntax, shipped with Sublime Text.

As all those aliases add a sub-scope to the used base scope, they need
to be added to this list here.

Example:

 All files ending with `.test.js` are scoped `source.js.unittest`.
 This commit makes sure to send the correct language id `javascript` to
 language-servers.