Implement client-initiated progress reporting (#1535)
For a selected number of requests we'll report progress in the status bar.
They are precisely the requests that are explicitly ran by the user. So for
instance: references, rename, symbols.

The implicit requests like code actions, highlights, hover, completions and
signature help are not progress-enabled.

Debounce by 200ms for fast requests.