Fix inconsistent popup's bottom margins (#1568)
ST can't margin-collapse margins of nested elements so elements with
bottom-margin that were direct children of the .lsp_popup element would
add an extra margin to the popup's bottom side.

Fix by instead always appending a spacer element with top margin at the
end of the popup contents. ST is able to margin-collapse margins of
sibling elements so that should create a consistent margin, regardless
of the content.

Fixed by creating a helper functions for showing and updating popups
that always add the spacer element. Those also have defaults for some
common arguments so that we don't have to always pass them.

One additional logic change, apart from adding the spacer element is
that we default to max_height=800 instead of the default max-height=240.

Resolves #1553