Handle moving views across windows
Basically the old state is cleared and we pretend we are dealing with a fresh
new ViewEventListener again. After thinking about this, it feels like ST should
do this as well: delete the old ViewEventListener, and create a fresh new
ViewEventListener for the new window. Because the new window may have other
settings applied to all views due to project settings, the ViewEventListener's
is_applicable classmethod may not be applicable anymore for the new window. In
any case, for now we'll have to use on_pre_move and on_post_move.