Replace document_highlight_scopes setting with better default scopes
This removes the user setting for the scopes used to color highlighted
regions of the textDocument/documentHighlight request. Users (or color
scheme authors) can configure their desired style via rules in the color
scheme instead. To easily allow fine-tuned colors, the more specific
scopes `markup.highlight.{text|read|write}.lsp` were added to the
corresponding DocumentHighlightKinds.

Additionally, the `Unknown` kind was removed, because it doesn't exist
in the specification. Complying with the specs, the default kind `Text`
is used now in case the highlight kind is omitted in the response.