Fix null session completions
Fixes #1622

The `sessions()` method returns a generator object, which causes the
query for `if not sessions` to always return false. As a result the
whole promise machinery is running even if no session is available.

This commit therefore queries `completion_promises` instead, to shorten
the route in case of no sessions being available.