Index SessionBuffer weak value dictionary by id(session) (#1762)
* Index SessionBuffer weak value dictionary by id(session)

Instead of the name. The config name caused weird behavior when
using the hover functionality. The text command of hover keeps the
session buffers alive. This ensures new session buffers are created
when doing a LSP: Restart Server. The CPython ID of session variable
is sufficiently unique, so we no longer need to key on the window ID.