Send a code lens request if one hasn't been sent yet
This is an extended fix related to #1759. The original intent of the
code was to allow code lenses to be filled without resorting to
reopening the file. However, the naive implementation caused an
infinite loop so it requires a separate sentinel value to differentiate
between an actual empty request vs an uninitialised request.