Even more fixes for matching up buffers with URIs
If the paths compare equal as strings, then we consider this a match, *even* if
the candidate file does not actually exist on-disk. This catches a case where
you delete a file on the file system but you have a view of it open in
Sublime Text. There is no callback in the API for that. And even so,
view.file_name() will still return the (now non-existing) file name.