Open existing views with custom URIs (#1779)
This doesn't actually work correctly for session buffers that get a different
session attached to them. For example, when you open a remote JSON schema with
the yaml-language-server, then the vscode-json-language-server is attached to
that virtual view. In that case this code fails.

We should instead have something like a per-window iteration over all session
buffers for all session. But this works okay for now for e.g. deno.