Fix inconsistency in indent adjustment behavior when completing
There is a small inconsistency when we're dealing with a text edit and the text
edit is not a snippet. In that case, if the new text of the completion
contained newlines, the subsequent line would not be "auto-indented". Whereas
with the three other possible cases, it would be auto-indented. That is because
the other three cases use the "insert" and "insert_snippet" commands, whereas
with this special case we were using the view.insert method.