Semantic highlighting (v2) (#1839)
This feature is enabled when setting the "semantic_highlighting" setting to true
in LSP.sublime-settings, and when the color scheme contains a certain scope.

The built-in color schemes are now overridden by this package so that they already
have the special scope targeted.

Both full and delta requests are implemented. The /refresh request from the server
to the client is also implemented.

The data is stored in the SessionBuffer.