remove phantom when they are clicked if they have textEdits
In order to follow the spec more precisly
/**
 * ...
 *
 * *Note* that edits are expected to change the document so that the inlay
 * hint (or its nearest variant) is now part of the document and the inlay
 * hint itself is **now obsolete**.
 *
 * ...
 */
textEdits?: TextEdit[];

Phantoms have an id but we cannot rely on that field because it is none at the time of creating the Phantom.
It is creted only when the phantom_set is created.

Instead attach a custom lsp_uuid field on the phantom