Handle custom URI schemes in hover text links (#2339)
* Handle custom URI schemes in hover text links

Before this change, custom URI schemes would be delegated to the browser.

With this change, an attempt is made to look for a session that can
handle the URI.

Moreover, if there is a fragment present in the URI, then it's assumed that
that fragment encodes the (row, col) to jump to. This logic is somewhat
dubious, but it's the only reasonable way (as far as I can see) to handle
this.

Anoher approach could be to introduce a new on_open_uri2_async callback
for plugins where they can also provide a (row, col) to jump to. But,
then we would also require an opt-in switch for plugins where they would
advertise that they can handle the "version 2" of on_open_uri.