Fix usage of sublime.score_selector
ST 4173 tweaks how selector scoring works resulting in LSP no longer
functioning.

As per the comment there seems to be a misunderstanding of how
`sublime.score_selector` works. The entire selector must always match
for a non-zero score, so if the goal is to check if it's not empty and
matches then checking for `>= 8` is simply wrong.