fix Discord link (#2449)
reported on discord

FichteFoll — Yesterday at 11:31 AM
FYI, I just noticed that you have a discord link to a channel in the footer of the docs (i.e. https://lsp.sublimetext.io/). This does not work for people not already part of the server. Please use the invite link in the channel topic instead (also mentioned in