Attempt at getting duplex pipes to work for Windows
No luck so far. I have checked out a development version of Node.js and compiled
it using the vcbuild.bat script. The command invocation that I used was:

    .\vcbuild.bat debug nosign x64

This builds all of V8, libuv and Node.js. The node.exe file is written to

    .\out\Debug\node.exe

A Visual Studio 2022 solution file is also created at .\node.sln. I have
this solution open and can attach to the node.exe with debug symbols after
first adding the following line just before the entrypoint in node_main.cc:

    while (!::IsDebuggerPresent()) {
        ::Sleep(100);
    }

This allows me to use CTRL+ALT+P to attach to the node.exe process when it
is spawned from Sublime Text as a subprocess. In the pipe_wrap.cc file, I have
put a breakpoint on the `PipeWrap::Open` method. In that method, the call to
`uv_pipe_open` fails. It fails because it is trying to open an invalid file
descriptor. The file descriptor is passed via the `NODE_CHANNEL_FD` environment
variable. This seems to go wrong and I don't understand why.

In order to use the duplex pipe transport, I modified my pyright JSON
configuration by adding these lines:

    "command": [
        "D:\\Development\\node\\out\\Debug\\node.exe",
        "${server_path}",
        "--node-ipc"
    ],
    "use_node_ipc": true,