make sure to restore the views in the original left-to-right order.
self.window.views() returns views ordered from left to right. When we iterate to close the views in the first loop, because closing a view shifts the indices of all subsequent views (to its right), we must iterate and grab the views from right to left (in reversed order). This prevents the shifting indices from causing errors when calling get_view_index for the remaining views.

When we restore the views, we do the opposite: we iterate from left to right (in original order).